# api/web/v1/web.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "web/v1/web.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "Web"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/web/v1/login": {
      "post": {
        "summary": "登陆",
        "operationId": "Web_Login",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiwebv1User"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1LoginReq"
            }
          }
        ],
        "tags": [
          "Web"
        ]
      }
    },
    "/web/v1/logout": {
      "post": {
        "summary": "登出",
        "operationId": "Web_Logout",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiwebv1LogoutReply"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/apiwebv1LogoutReq"
            }
          }
        ],
        "tags": [
          "Web"
        ]
      }
    },
    "/web/v1/users": {
      "get": {
        "summary": "用户列表",
        "operationId": "Web_ListUser",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiwebv1ListUserReply"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "Web"
        ]
      }
    }
  },
  "definitions": {
    "apiwebv1ListUserReply": {
      "type": "object",
      "properties": {
        "items": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/apiwebv1User"
          }
        },
        "total": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "apiwebv1LogoutReply": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "uint64"
        }
      },
      "title": "回应 - 登出"
    },
    "apiwebv1LogoutReq": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "uint64"
        }
      },
      "title": "请求 - 登出"
    },
    "apiwebv1User": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "uint64"
        },
        "userName": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "nickName": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "authoritys": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "token": {
          "type": "string"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "typeUrl": {
          "type": "string"
        },
        "value": {
          "type": "string",
          "format": "byte"
        }
      }
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "v1LoginReq": {
      "type": "object",
      "properties": {
        "userName": {
          "type": "string",
          "title": "用户名"
        },
        "password": {
          "type": "string"
        }
      },
      "title": "请求 - 登录"
    }
  }
}